Updated bounds, added scaling to fit
I realized the coordinate issue (the 0.15-inch offset) I had earlier was
because the text origin is the bottom of the line of text--not the top.
(12-point type is 0.167 inches high, hence the problem.)

The bounds for the line with the SSN/FRN field have been shifted a bit in order
to accommodate the second line of the SSN/FRN field description.

Font size now gets scaled in order to make the line fit if it's too wide for
the given bounds.